ci: clippy fixes

+++ src/util/bip32.rs
@@ -637,7 +637,7 @@ impl ExtendedPrivKey {
             Network::Bitcoin => [0x04, 0x88, 0xAD, 0xE4],
             Network::Testnet | Network::Signet | Network::Regtest => [0x04, 0x35, 0x83, 0x94],
         }[..]);
-        ret[4] = self.depth as u8;
+        ret[4] = self.depth;
         ret[5..9].copy_from_slice(&self.parent_fingerprint[..]);
         ret[9..13].copy_from_slice(&endian::u32_to_array_be(u32::from(self.child_number)));
         ret[13..45].copy_from_slice(&self.chain_code[..]);
@@ -775,7 +775,7 @@ impl ExtendedPubKey {
             Network::Bitcoin => [0x04u8, 0x88, 0xB2, 0x1E],
             Network::Testnet | Network::Signet | Network::Regtest => [0x04u8, 0x35, 0x87, 0xCF],
         }[..]);
-        ret[4] = self.depth as u8;
+        ret[4] = self.depth;
         ret[5..9].copy_from_slice(&self.parent_fingerprint[..]);
         ret[9..13].copy_from_slice(&endian::u32_to_array_be(u32::from(self.child_number)));
         ret[13..45].copy_from_slice(&self.chain_code[..]);
